pyproject.toml
[project]
name = "abyss"
version = "2026.04.30"
description = "Telegram + Claude Code personal AI assistant"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "croniter>=2.0",
    "pyzipper>=0.3.6",
    "claude-agent-sdk>=0.1.48",
    "aiohttp>=3.9",
]

[project.scripts]
abyss = "abyss.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/abyss"]

[tool.hatch.build.targets.wheel.force-include]
"abysscope/src" = "abyss/abysscope_data/src"
"abysscope/public" = "abyss/abysscope_data/public"
"abysscope/package.json" = "abyss/abysscope_data/package.json"
"abysscope/package-lock.json" = "abyss/abysscope_data/package-lock.json"
"abysscope/tsconfig.json" = "abyss/abysscope_data/tsconfig.json"
"abysscope/next.config.ts" = "abyss/abysscope_data/next.config.ts"
"abysscope/postcss.config.mjs" = "abyss/abysscope_data/postcss.config.mjs"
"abysscope/components.json" = "abyss/abysscope_data/components.json"
"abysscope/eslint.config.mjs" = "abyss/abysscope_data/eslint.config.mjs"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
markers = ["evaluation: real Claude API calls for parsing quality evaluation"]
addopts = "--ignore=tests/evaluation"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
